changed default editor to nvim.
fixed mapping for handling merge conflicts and went over help docs. basically
get a 3 way split. on the left is what you have locally. the right is what is
comming in and middle is what will be kept.

so use c[ and c] to jump between hunks and if you want whats on the left push
d2o if you want what is on the right push d3o (may remap that later haven't
decided)

when done hit :only or <Ctrl-w>o (same thing) to close all but the middle file.
save it test it push it or move to next file.